embedding-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
pydantic==2.10.4
python-multipart==0.0.20
numpy==1.26.4
requests==2.31.0

